moved MovieService.java file into services package with other services and deleted the duplicated service package
