+++ movie-planner/src/main/java/com/sparta/movieplanner/controllers/web/MovieController.java
@@ -1,6 +1,6 @@
 package com.sparta.movieplanner.controllers.web;
 
-import com.sparta.movieplanner.service.MovieService;
+import com.sparta.movieplanner.services.MovieService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
